requirements.txt
Flask==3.0.2
Pillow==10.2.0
imageio==2.33.1
python-dotenv==1.0.1
numpy==1.26.3 